nextjs-fastapi/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
google-generativeai==0.5.0
python-dotenv==1.0.1